administrador/requirements.txt
asgiref==3.8.1
attrs==25.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
et_xmlfile==2.0.0
future==1.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
numpy==2.2.6
packaging==25.0
Pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==25.0.0
python-decouple==3.8
pytz==2025.2
redis==5.2.1
serializers==0.2.4
service-identity==24.2.0
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.6.0
xlrd==2.0.1
zope.interface==7.2
gunicorn==21.2.0
ariadne==0.26.2
django-stubs==5.2.8
djangorestframework-stubs==3.16.6
mypy==1.19.0
ruff==0.14.8
